netsec-cloud/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netsec-cloud"
version = "0.1.0"
description = "Cloud Security Scanner - Multi-cloud security scanning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "NetSec Toolkit Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Networking",
]
dependencies = [
    "boto3>=1.28.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
azure = [
    "azure-identity>=1.15.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-storage>=21.0.0",
    "azure-mgmt-network>=24.0.0",
]
gcp = [
    "google-cloud-resource-manager>=1.10.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-compute>=1.14.0",
]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
netsec-cloud = "netsec_cloud.cli.main:cli"

[tool.setuptools]
packages = ["netsec_cloud"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
